Separated logic or store tracking
